Move shared files to pythonshared
The pythonshared dir should now be in PYTHONPATH for the
Monitor and Updater, these common scripts can be moved there.
